+++ apps/antalmanac/src/components/Header/halloween-logo.png
