+++ apps/antalmanac/src/components/Header/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Header';
