+++ apps/antalmanac/src/routes/Home.tsx
@@ -3,12 +3,12 @@ import DateFnsUtils from '@date-io/date-fns';
 import { MuiPickersUtilsProvider } from '@material-ui/pickers';
 import { Box, CssBaseline, useMediaQuery, useTheme } from '@material-ui/core';
 
-import Bar from '$components/AppBar/CustomAppBar';
+import Header from '$components/Header';
 import MobileHome from '$components/MobileHome';
 import PatchNotes from '$components/PatchNotes';
 import Calendar from '$components/Calendar/CalendarRoot';
 import DesktopTabs from '$components/RightPane/RightPaneRoot';
-import NotificationSnackbar from '$components/AppBar/NotificationSnackbar';
+import NotificationSnackbar from '$components/NotificationSnackbar';
 
 export default function Home() {
     const isMobileScreen = useMediaQuery('(max-width: 750px)');
@@ -18,7 +18,7 @@ export default function Home() {
         <MuiPickersUtilsProvider utils={DateFnsUtils}>
             <CssBaseline />
             <PatchNotes />
-            <Bar />
+            <Header />
             {isMobileScreen ? (
                 <MobileHome />
             ) : (
